Fixed tests, dropped < py3.5 support
Closes #59

Reverted the setup.py change as aiobotocore specifies a specific version of botocore, which isn't the absolute latest which will then complain when the latest boto3 is installed.

Removing the yield from breaks some edge cases so instead I just duped the function, made it a coroutine and called it async as it was already being called from a coro.