Make DEFAULT_SESSION use be thread-safe
I have an implementation that uses threads to distribute
calls to S3 resources. When using several threads at once,
lots of errors emerged.

I figured out that it was because of this:

https://boto3.amazonaws.com/v1/documentation/api/latest/guide/session.html#multithreading-or-multiprocessing-with-sessions

Meaning that Sessions can't be reused by different threads.

To solve this what I did was to change the way we use the global
variable DEFAULT_SESSION to make it possible to have one default
session for each thread.